requirements.txt
absl-py==2.1.0
accelerate==1.5.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
art==6.4
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
distro==1.9.0
filelock==3.13.1
fonttools==4.56.0
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
grpcio==1.71.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
Jinja2==3.1.4
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
langchain==0.3.21
langchain-core==0.3.47
langchain-openai==0.3.9
langchain-text-splitters==0.3.7
langgraph==0.3.18
langgraph-checkpoint==2.0.21
langgraph-prebuilt==0.1.4
langgraph-sdk==0.1.58
langsmith==0.3.18
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.8.4
mpmath==1.3.0
msgpack==1.1.0
narwhals==1.31.0
networkx==3.3
numpy==1.24.0
openai==1.68.0
opencv-python==4.9.0.80
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pillow==10.2.0
protobuf==5.29.4
psutil==7.0.0
pyarrow==19.0.1
pycm==4.2
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.23.1
safetensors==0.5.3
scikit-learn==1.4.2
scikit-multilearn==0.2.0
scipy==1.15.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.1
SQLAlchemy==2.0.39
streamlit==1.43.2
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tiktoken==0.9.0
timm==0.9.16
tokenizers==0.21.1
toml==0.10.2
torch==2.4.1+cu121
torchaudio==2.4.1+cu121
torchvision==0.19.1+cu121
tornado==6.4.2
tqdm==4.67.1
transformers==4.47.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
watchdog==6.0.0
Werkzeug==3.1.3
zstandard==0.23.0
